Add support for 'cleanpath' option
Unload command in Redshift supports, among others, the 'cleanup' option, which hasn't been supported by this macro so far. This change is aimed at adding the missing support. 

Note that "cleanpath" cannot be used together with "allowoverwrite"; no check is made in (this version of) the macro to test for their co-presence.